+++ extensions/youtube/src/main/java/app/revanced/extension/youtube/patches/components/HideInfoCardsFilter.java
@@ -3,9 +3,9 @@ package app.revanced.extension.youtube.patches.components;
 import app.revanced.extension.youtube.settings.Settings;
 
 @SuppressWarnings("unused")
-public final class HideInfoCardsFilterPatch extends Filter {
+public final class HideInfoCardsFilter extends Filter {
 
-    public HideInfoCardsFilterPatch() {
+    public HideInfoCardsFilter() {
         addIdentifierCallbacks(
                 new StringFilterGroup(
                         Settings.HIDE_INFO_CARDS,